+++ docs/src/.vitepress/config.ts
@@ -82,7 +82,7 @@ export default defineConfig({
             link: 'https://www.npmjs.com/package/@mcp-ui/server',
           },
           {
-            text: 'Ruby Gem',
+            text: 'mcp_ui_server Gem',
             link: 'https://rubygems.org/gems/mcp_ui_server',
           },
         ],
@@ -97,6 +97,7 @@ export default defineConfig({
             { text: 'Introduction', link: '/guide/introduction' },
             { text: 'Installation', link: '/guide/getting-started' },
             { text: 'Core Concepts', link: '/guide/protocol-details' },
+            { text: 'Embeddable UI', link: '/guide/embeddable-ui' },
           ],
         },
         {
@@ -136,24 +137,23 @@ export default defineConfig({
                   link: '/guide/client/resource-renderer',
                 },
                 {
-                  text: 'Custom Component Libraries',
-                  link: '/guide/client/custom-component-libraries',
+                  text: 'React Usage & Examples',
+                  link: '/guide/client/react-usage-examples'
                 },
                 {
-                  text: 'Using a Proxy',
-                  link: '/guide/client/using-a-proxy',
-                },
-                {
-                  text: 'HTMLResourceRenderer',
-                  link: '/guide/client/html-resource',
-                },
-                {
-                  text: 'RemoteDOMResourceRenderer',
-                  link: '/guide/client/remote-dom-resource',
-                },
+                  text: 'Web Component Usage & Examples',
+                  link: '/guide/client/wc-usage-examples',
+                }
               ],
             },
-            { text: 'Usage & Examples', link: '/guide/client/usage-examples' },
+            {
+              text: 'Custom Component Libraries',
+              link: '/guide/client/custom-component-libraries',
+            },
+            {
+              text: 'Using a Proxy',
+              link: '/guide/client/using-a-proxy',
+            }
           ],
         },
       ],
@@ -193,8 +193,14 @@ export default defineConfig({
 
     socialLinks: [
       { icon: 'github', link: 'https://github.com/idosal/mcp-ui' },
+      { icon: 'discord', link: 'https://discord.gg/CEAG4KW7ZH' },
       { icon: 'npm', link: 'https://www.npmjs.com/package/@mcp-ui/server' },
-      { icon: 'github', link: 'https://rubygems.org/gems/mcp_ui_server' },
+      {
+        icon: {
+          svg: '<svg viewBox="0 0 256 293" xmlns="http://www.w3.org/2000/svg" preserveAspectRatio="xMinYMin meet"><g fill="currentColor"><path d="M76.748 97.434l-.163-.163-36.11 36.11 87.674 87.512 36.11-35.948 51.564-51.563-36.11-36.11v-.164H76.584l.163.326z"/><path d="M127.823.976L.135 74.173v146.395l127.688 73.197 127.689-73.197V74.173L127.823.976zm103.29 205.603l-103.29 59.534-103.29-59.534V87.837l103.29-59.534 103.29 59.534v118.742z"/></g></svg>',
+        },
+        link: 'https://rubygems.org/gems/mcp_ui_server',
+      },
     ],
 
     footer: {