+++ sdks/typescript/client/CHANGELOG.md
@@ -1,3 +1,38 @@
+## [5.6.2](https://github.com/idosal/mcp-ui/compare/v5.6.1...v5.6.2) (2025-08-05)
+
+
+### Bug Fixes
+
+* bump client version ([75c9236](https://github.com/idosal/mcp-ui/commit/75c923689654b4443ad1093fafc0bad16902e4cc))
+
+## [5.6.1](https://github.com/idosal/mcp-ui/compare/v5.6.0...v5.6.1) (2025-08-05)
+
+
+### Bug Fixes
+
+* wc dist overwrite ([#63](https://github.com/idosal/mcp-ui/issues/63)) ([9e46c56](https://github.com/idosal/mcp-ui/commit/9e46c56c7a8908410fad6d08a5d845139e93f80f))
+
+# [5.6.0](https://github.com/idosal/mcp-ui/compare/v5.5.0...v5.6.0) (2025-08-03)
+
+
+### Features
+
+* add UIResourceRenderer Web Component ([#58](https://github.com/idosal/mcp-ui/issues/58)) ([ec8f299](https://github.com/idosal/mcp-ui/commit/ec8f2994ecf36774e6ad5191654ba22946d0ee49))
+
+# [5.5.0](https://github.com/idosal/mcp-ui/compare/v5.4.0...v5.5.0) (2025-08-02)
+
+
+### Features
+
+* auto resize with the autoResizeIframe prop ([#56](https://github.com/idosal/mcp-ui/issues/56)) ([76c867a](https://github.com/idosal/mcp-ui/commit/76c867a569b72aed892290aa84e1194ab8eb79ce))
+
+# [5.4.0](https://github.com/idosal/mcp-ui/compare/v5.3.1...v5.4.0) (2025-08-01)
+
+
+### Features
+
+* send render data to the iframe ([#51](https://github.com/idosal/mcp-ui/issues/51)) ([d38cfc7](https://github.com/idosal/mcp-ui/commit/d38cfc7925061c1ae1911bdee408033c8e9f283d))
+
 ## [5.3.1](https://github.com/idosal/mcp-ui/compare/v5.3.0...v5.3.1) (2025-07-19)